% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_selection.R
\name{analyze_structure_diversity}
\alias{analyze_structure_diversity}
\title{Analyze Structure Diversity}
\usage{
analyze_structure_diversity(structures)
}
\arguments{
\item{structures}{List of TreeStructure objects}
}
\value{
List with:
  \item{n_unique}{Integer: number of unique structures}
  \item{shannon_entropy}{Numeric: Shannon entropy of structure distribution}
  \item{simpson_index}{Numeric: Simpson diversity index}
  \item{modal_frequency}{Numeric: frequency of modal structure}
}
\description{
Compute summary statistics for structure diversity across a set of fitted structures.
}
\details{
Diversity metrics:
- n_unique: Higher = more diverse
- shannon_entropy: Higher = more diverse (0 = all same, log(M) = uniform)
- simpson_index: Higher = more diverse (0 = all same, 1-1/M = uniform)
- modal_frequency: Lower = more diverse
}
\examples{
\dontrun{
diversity <- analyze_structure_diversity(structures)
cat(sprintf("Unique structures: \%d / \%d\n",
            diversity$n_unique, length(structures)))
}

}
