% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon_analysis.R
\name{compare_rashomon_sets}
\alias{compare_rashomon_sets}
\title{Compare Rashomon Sets}
\usage{
compare_rashomon_sets(set1, set2, verbose = FALSE)
}
\arguments{
\item{set1}{First Rashomon set (list of tree JSON strings)}

\item{set2}{Second Rashomon set (list of tree JSON strings)}

\item{verbose}{Logical: whether to print detailed comparison}
}
\value{
A list containing:
  \item{common_trees}{Trees appearing in both sets}
  \item{unique_to_set1}{Trees only in set1}
  \item{unique_to_set2}{Trees only in set2}
  \item{set1_size}{Size of set1}
  \item{set2_size}{Size of set2}
  \item{common_size}{Number of common trees}
}
\description{
Compare two Rashomon sets to find common and unique trees.
}
\examples{
\dontrun{
# Compare two Rashomon sets
comparison <- compare_rashomon_sets(set1, set2, verbose = TRUE)
print(comparison)
}

}
