% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon.R
\name{compare_trees}
\alias{compare_trees}
\title{Compare Two Trees for Structural Equality}
\usage{
compare_trees(tree1, tree2)
}
\arguments{
\item{tree1}{First TreeClassifier object}

\item{tree2}{Second TreeClassifier object}
}
\value{
Logical: TRUE if trees have identical structure, FALSE otherwise
}
\description{
Check if two tree structures are identical by comparing their JSON representations.
}
\examples{
\dontrun{
trees <- get_rashomon_trees(model)
are_same <- compare_trees(trees[[1]], trees[[2]])
}

}
