% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_regularization.R
\name{compute_safe_model_limit}
\alias{compute_safe_model_limit}
\title{Compute safe model_limit for given lambda and n}
\usage{
compute_safe_model_limit(lambda, n, p_features)
}
\arguments{
\item{lambda}{Regularization value}

\item{n}{Sample size}

\item{p_features}{Number of binary features (after discretization)}
}
\value{
Safe model_limit value
}
\description{
Theory prescribes lambda ~ log(n)/n, but small lambda requires
larger model_limit for solver to complete successfully.
}
\keyword{internal}
