% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{compute_thresholds}
\alias{compute_thresholds}
\title{Compute Discretization Thresholds}
\usage{
compute_thresholds(x, method, n_bins, user_threshold = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{method}{Discretization method}

\item{n_bins}{Number of bins for quantiles}

\item{user_threshold}{Optional user-provided thresholds}
}
\value{
Numeric vector of thresholds
}
\description{
Compute Discretization Thresholds
}
\keyword{internal}
