% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{coord_tree_count_leaves}
\alias{coord_tree_count_leaves}
\title{Number of leaves in a coordinate-space (sub)tree}
\usage{
coord_tree_count_leaves(node)
}
\description{
Number of leaves in a coordinate-space (sub)tree
}
\keyword{internal}
