% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{coord_tree_node_at}
\alias{coord_tree_node_at}
\title{Get the node at `path` (see [coord_tree_split_paths()] for the path format)}
\usage{
coord_tree_node_at(tree, path)
}
\description{
Get the node at `path` (see [coord_tree_split_paths()] for the path format)
}
\keyword{internal}
