% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{coord_tree_set_node_at}
\alias{coord_tree_set_node_at}
\title{Return `tree` with the node at `path` replaced by `value`}
\usage{
coord_tree_set_node_at(tree, path, value)
}
\description{
Return `tree` with the node at `path` replaced by `value`
}
\keyword{internal}
