% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{coord_tree_training_sse}
\alias{coord_tree_training_sse}
\title{Total training risk (SSE) of a coordinate-space tree}
\usage{
coord_tree_training_sse(tree, X, y)
}
\description{
Total training risk (SSE) of a coordinate-space tree
}
\keyword{internal}
