% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon.R
\name{count_trees}
\alias{count_trees}
\title{Get Number of Trees in Rashomon Set}
\usage{
count_trees(model)
}
\arguments{
\item{model}{A trained TreeFARMS model object (class `treefarms_model`)}
}
\value{
Integer: number of trees in the Rashomon set
}
\description{
Get the count of trees in a model's Rashomon set.
}
\examples{
\dontrun{
model <- treefarms(X, y, regularization = 0.1)
n_trees <- count_trees(model)
}

}
