% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{count_trees_in_result}
\alias{count_trees_in_result}
\title{Count Trees in Rashomon Set}
\usage{
count_trees_in_result(result_data, tree_json)
}
\arguments{
\item{result_data}{Parsed JSON data from C++}

\item{tree_json}{Extracted tree structure}
}
\value{
Integer count of trees in Rashomon set
}
\description{
Count Trees in Rashomon Set
}
\keyword{internal}
