% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_fitted_rashomon.R
\name{create_stratified_folds_from_y}
\alias{create_stratified_folds_from_y}
\title{Create Stratified Folds}
\usage{
create_stratified_folds_from_y(y, K)
}
\arguments{
\item{y}{Binary outcome vector}

\item{K}{Number of folds}
}
\value{
List of length K, each element containing indices for that fold
}
\description{
Internal function to create K stratified folds based on outcome variable.
}
\keyword{internal}
