% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{determine_node_type}
\alias{determine_node_type}
\title{Determine Node Type}
\usage{
determine_node_type(node_data)
}
\arguments{
\item{node_data}{List: node data}
}
\value{
Character string: node type
}
\description{
Determine the type of a tree node.
}
\keyword{internal}
