% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_optimality_certificate.R
\name{.certificate_result}
\alias{.certificate_result}
\title{Assemble the plain-list return value of [certify_local_optimality()]}
\usage{
.certificate_result(
  certified,
  reason,
  margin,
  margin_rel,
  objective,
  sse,
  n_leaves,
  lambda_sse,
  tol,
  worst,
  by_kind,
  detail,
  notes
)
}
\description{
Assemble the plain-list return value of [certify_local_optimality()]
}
\keyword{internal}
