% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{data}
\name{.CLASSIFICATION_THRESHOLD}
\alias{.CLASSIFICATION_THRESHOLD}
\title{S3 Predict Methods for TreeFARMS Models}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.CLASSIFICATION_THRESHOLD
}
\description{
S3 methods for making predictions with TreeFARMS models.
Provides a unified interface for all model types.
}
\keyword{datasets}
