% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_loss.R
\name{example_custom_absolute_error}
\alias{example_custom_absolute_error}
\title{Create Custom Absolute Error Loss (Example)}
\usage{
example_custom_absolute_error()
}
\value{
A list with the required custom loss components
}
\description{
Example custom loss function implementing absolute error (L1) loss.
}
\examples{
\dontrun{
custom_l1 <- example_custom_absolute_error()
m <- optimaltrees(X, y, loss_function = "custom", custom_loss = custom_l1)
}

}
