% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_loss.R
\name{example_custom_squared_error}
\alias{example_custom_squared_error}
\title{Create Custom Squared Error Loss (Example)}
\usage{
example_custom_squared_error()
}
\value{
A list with the required custom loss components
}
\description{
Example custom loss function implementing squared error loss.
Provided as a template for users to create their own custom losses.
}
\examples{
\dontrun{
# Use the example custom loss
custom_se <- example_custom_squared_error()
m <- optimaltrees(X, y, loss_function = "custom", custom_loss = custom_se)
}

}
