% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{finalize_result_object}
\alias{finalize_result_object}
\title{Finalize Result Object with Metadata}
\usage{
finalize_result_object(
  result,
  model_obj,
  X,
  y,
  store_training_data,
  discretization_metadata,
  X_original
)
}
\arguments{
\item{result}{Result list to finalize}

\item{model_obj}{Model object to add}

\item{X}{Training features}

\item{y}{Training labels}

\item{store_training_data}{Logical. Whether training data should be stored}

\item{discretization_metadata}{Discretization metadata}

\item{X_original}{Original feature names}
}
\value{
Finalized result object with class
}
\description{
Finalize Result Object with Metadata
}
\keyword{internal}
