% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_refit.R
\name{find_split_for_path}
\alias{find_split_for_path}
\title{Find Split for Given Path}
\usage{
find_split_for_path(all_splits, target_path)
}
\arguments{
\item{all_splits}{List of split information}

\item{target_path}{Integer vector (path from root)}
}
\value{
Split information, or stops with error if not found
}
\description{
Find Split for Given Path
}
\keyword{internal}
