% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Get accuracy from a treefarms model (with lazy evaluation)}
\usage{
get_accuracy(object)
}
\arguments{
\item{object}{A optimaltrees_model object}
}
\value{
Training accuracy (numeric)
}
\description{
Get accuracy from a treefarms model (with lazy evaluation)
}
