% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{get_edge_label}
\alias{get_edge_label}
\title{Get Edge Label}
\usage{
get_edge_label(node_data)
}
\arguments{
\item{node_data}{List: node data}
}
\value{
Character string: edge label
}
\description{
Get the label for an edge.
}
\keyword{internal}
