% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{get_node_loss}
\alias{get_node_loss}
\title{Get Node Loss}
\usage{
get_node_loss(node_data)
}
\arguments{
\item{node_data}{List: node data}
}
\value{
Numeric: loss value
}
\description{
Get the loss value for a node.
}
\keyword{internal}
