% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{get_node_prediction}
\alias{get_node_prediction}
\title{Get Node Prediction}
\usage{
get_node_prediction(node_data)
}
\arguments{
\item{node_data}{List: node data}
}
\value{
Numeric: prediction value
}
\description{
Get the prediction value for a node.
}
\keyword{internal}
