% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{get_node_threshold}
\alias{get_node_threshold}
\title{Get Node Threshold}
\usage{
get_node_threshold(node_data)
}
\arguments{
\item{node_data}{List: node data}
}
\value{
Numeric: threshold value
}
\description{
Get the threshold value for a decision node.
}
\keyword{internal}
