% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions from a treefarms model (with lazy evaluation)}
\usage{
get_predictions(object)
}
\arguments{
\item{object}{A optimaltrees_model object}
}
\value{
A vector of binary predictions (0/1)
}
\description{
Get predictions from a treefarms model (with lazy evaluation)
}
