% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon_utils.R
\name{get_rashomon_prop}
\alias{get_rashomon_prop}
\title{Get property from CFRashomon object (S7 or S3)}
\usage{
get_rashomon_prop(obj, prop)
}
\description{
Get property from CFRashomon object (S7 or S3)
}
\keyword{internal}
