% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{is_binary}
\alias{is_binary}
\title{Check if a Vector is Binary}
\usage{
is_binary(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Logical: TRUE if all non-NA values are in {0, 1}
}
\description{
Check if a Vector is Binary
}
\keyword{internal}
