% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{leaf_assignments}
\alias{leaf_assignments}
\title{Leaf assigned to each row of `newdata` by a `RefinedTreeModel`}
\usage{
leaf_assignments(object, newdata)
}
\arguments{
\item{object}{A [RefinedTreeModel].}

\item{newdata}{Data.frame or matrix with the coordinates \code{object}
splits on (\code{object@coords}), as raw (non-discretized) values.}
}
\value{
Integer vector of length \code{nrow(newdata)}: the leaf `id`
  (see \code{\link{stage_b}}'s node schema) each row falls into.
}
\description{
Leaf assigned to each row of `newdata` by a `RefinedTreeModel`
}
