% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_b.R
\name{n_leaves}
\alias{n_leaves}
\title{Number of leaves in a `RefinedTreeModel`}
\usage{
n_leaves(object)
}
\arguments{
\item{object}{A [RefinedTreeModel].}
}
\value{
Integer.
}
\description{
Number of leaves in a `RefinedTreeModel`
}
