% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{OptimalTreesModelLoaded}
\alias{OptimalTreesModelLoaded}
\title{Loaded Optimal Trees Model}
\usage{
OptimalTreesModelLoaded(
  model = (structure(function (loss_function = character(0), regularization = numeric(0),
    n_trees = integer(0), trees = list(), accuracy = numeric(0), predictions = NULL,
    probabilities = NULL, X_train = NULL, y_train = NULL, discretization_metadata =
    list(), is_regression = logical(0)) 
 {
     loss_function
     regularization
    
    n_trees
     trees
     accuracy
     predictions
     probabilities
     X_train
   
     y_train
     discretization_metadata
     is_regression
     new_object(S7_object(),
    loss_function = loss_function, regularization = regularization, n_trees = n_trees,
    trees = trees, accuracy = accuracy, predictions = predictions, probabilities =
    probabilities, X_train = X_train, y_train = y_train, discretization_metadata =
    discretization_metadata, is_regression = is_regression)
 }, name =
    "OptimalTreesModel", parent = structure(function () 
 {
     .Call(S7_object_)
 },
    name = "S7_object", properties = list(), abstract = FALSE, constructor = function ()
    
 {
     .Call(S7_object_)
 }, validator = function (self) 
 {
     if
    (!is_S7_type(self)) {
         "Underlying data is corrupt"
     }
 }, class =
    c("S7_class", "S7_object")), properties = list(loss_function = structure(list(name =
    "loss_function", class = structure(list(class = "character", constructor_name =
    "character", constructor = function (.data = character(0)) 
 .data, validator =
    function (object) 
 {
     if (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property"), regularization = structure(list(name =
    "regularization", class = structure(list(class = "double", constructor_name =
    "double", constructor = function (.data = numeric(0)) 
 .data, validator = function
    (object) 
 {
     if (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property"), n_trees = structure(list(name = "n_trees", class =
    structure(list(class = "integer", constructor_name = "integer", constructor =
    function (.data = integer(0)) 
 .data, validator = function (object) 
 {
     if
    (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property"), trees = structure(list(name = "trees", class =
    structure(list(class = "list", constructor_name = "list", constructor = function
    (.data = list()) 
 .data, validator = function (object) 
 {
     if
    (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property"), accuracy = structure(list(name = "accuracy", class =
    structure(list(classes = list(structure(list(class = "double", constructor_name =
    "double", constructor = function (.data = numeric(0)) 
 .data, validator = function
    (object) 
 {
     if (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), structure(list(class = "logical", constructor_name =
    "logical", constructor = function (.data = logical(0)) 
 .data, validator = function
    (object) 
 {
     if (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"))), class = "S7_union"), getter = NULL, setter = NULL,
    validator = NULL, default = NULL), class = "S7_property"), predictions =
    structure(list(name = "predictions", class = structure(list(), class = "S7_any"),
    getter = NULL, setter = NULL, validator = NULL, default = NULL), class =
    "S7_property"), probabilities = structure(list(name = "probabilities", class =
    structure(list(), class = "S7_any"), getter = NULL, setter = NULL, validator = NULL,
    default = NULL), class = "S7_property"), 
     X_train = structure(list(name =
    "X_train", class = structure(list(), class = "S7_any"), getter = NULL, setter = NULL,
    validator = NULL, default = NULL), class = "S7_property"), y_train =
    structure(list(name = "y_train", class = structure(list(), class = "S7_any"), getter
    = NULL, setter = NULL, validator = NULL, default = NULL), class = "S7_property"),
    discretization_metadata = structure(list(name = "discretization_metadata", class =
    structure(list(class = "list", constructor_name = "list", constructor = function
    (.data = list()) 
     .data, validator = function (object) 
     {
         if
    (base_class(object) != name) {
            
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
        
    }
     }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL,
    default = NULL), class = "S7_property"), is_regression = structure(list(name =
    "is_regression", class = structure(list(class = "logical", constructor_name =
    "logical", constructor = function (.data = logical(0)) 
     .data, validator =
    function (object) 
     {
         if (base_class(object) != name) {
            
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
        
    }
     }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL,
    default = NULL), class = "S7_property")), abstract = FALSE, constructor = function
    (loss_function = character(0), regularization = numeric(0), n_trees = integer(0),
    trees = list(), accuracy = numeric(0), predictions = NULL, probabilities = NULL,
    X_train = NULL, y_train = NULL, discretization_metadata = list(), is_regression =
    logical(0)) 
 {
     loss_function
     regularization
     n_trees
     trees
    
    accuracy
     predictions
     probabilities
     X_train
     y_train
    
    discretization_metadata
     is_regression
     new_object(S7_object(), loss_function
    = loss_function, regularization = regularization, n_trees = n_trees, trees = trees,
    accuracy = accuracy, predictions = predictions, probabilities = probabilities,
    X_train = X_train, y_train = y_train, discretization_metadata =
    discretization_metadata, is_regression = is_regression)
 }, validator = function
    (self) 
 {
     if (self@regularization <= 0) {
        
    return("@regularization must be positive")
     }
     if (self@n_trees < 0) {
      
      return("@n_trees must be non-negative")
     }
     if (length(self@trees) !=
    self@n_trees) {
         return("length(@trees) must equal @n_trees")
     }
     if
    (!self@is_regression && !is.na(self@accuracy)) {
         if (self@accuracy < 0 ||
    self@accuracy > 1) {
             return("@accuracy must be in [0, 1]")
         }
  
      }
     valid_losses <- c("misclassification", "log_loss", "squared_error",
    "absolute_error", "huber", "quantile", "custom")
     if (!self@loss_function \%in\%
    valid_losses) {
         return(sprintf("@loss_function must be one of: \%s",
    paste(valid_losses, collapse = ", ")))
     }
     if (!is.null(self@X_train)) {
    
        if (!is.data.frame(self@X_train) && !is.matrix(self@X_train)) {
            
    return("@X_train must be a data.frame or matrix")
         }
     }
     if
    (!is.null(self@y_train)) {
         if (!is.numeric(self@y_train) &&
    !is.integer(self@y_train)) {
            
    return("@y_train must be numeric or integer")
         }
     }
     if
    (!is.null(self@X_train) && !is.null(self@y_train)) {
         n_train <-
    nrow(self@X_train)
         n_outcome <- length(self@y_train)
         if (n_train !=
    n_outcome) {
            
    return(sprintf("nrow(@X_train) = \%d must equal length(@y_train) = \%d", n_train,
    n_outcome))
         }
     }
 }, class = c("S7_class", "S7_object")))(),
  loaded_from = "",
  loaded_at = structure(1774463691.59901, class = c("POSIXct", "POSIXt"))
)
}
\description{
Wrapper for models loaded from disk via save_optimaltrees/load_optimaltrees.
Contains the model plus serialization metadata.
}
\details{
This class exists to distinguish loaded models from freshly trained ones,
which can be useful for deciding whether certain operations (like retraining)
are safe.
}
