% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.optimaltrees_model}
\alias{plot.optimaltrees_model}
\title{Plot method for optimaltrees_model}
\usage{
\method{plot}{optimaltrees_model}(x, tree_index = 1, ...)
}
\arguments{
\item{x}{A optimaltrees_model object}

\item{tree_index}{Integer: which tree to plot (default: 1)}

\item{...}{Additional arguments (unused)}
}
\description{
Plot method for optimaltrees_model
}
