% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.treefarms_model}
\alias{plot.treefarms_model}
\title{Plot method for treefarms_model}
\usage{
\method{plot}{treefarms_model}(x, tree_index = 1, ...)
}
\arguments{
\item{x}{A treefarms_model object}

\item{tree_index}{Integer: which tree to plot (default: 1)}

\item{...}{Additional arguments (unused)}
}
\description{
Plot method for treefarms_model
}
