% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_cf_overview}
\alias{plot_cf_overview}
\title{Plot Cross-Fitted Rashomon Overview}
\usage{
plot_cf_overview(cf_result, ...)
}
\arguments{
\item{cf_result}{A cf_rashomon object}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create an overview visualization of cross-fitted Rashomon results.
}
