% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_cf_stability}
\alias{plot_cf_stability}
\title{Plot Cross-Fitted Rashomon Stability}
\usage{
plot_cf_stability(cf_result, ...)
}
\arguments{
\item{cf_result}{A cf_rashomon object}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a stability visualization for cross-fitted Rashomon results.
}
