% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_cf_trees}
\alias{plot_cf_trees}
\title{Plot Cross-Fitted Rashomon Trees}
\usage{
plot_cf_trees(cf_result, interactive = FALSE, ...)
}
\arguments{
\item{cf_result}{A cf_rashomon object}

\item{interactive}{Logical: whether to create interactive plot}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object or interactive plot
}
\description{
Create visualizations of trees in cross-fitted Rashomon results.
}
