% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_tree_comparison}
\alias{plot_tree_comparison}
\title{Plot Tree Comparison}
\usage{
plot_tree_comparison(model, trees_to_compare = NULL, ...)
}
\arguments{
\item{model}{A treefarms model object}

\item{trees_to_compare}{Integer vector: which trees to compare}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a comparison visualization of multiple trees.
}
