% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_tree_interactive}
\alias{plot_tree_interactive}
\title{Plot Tree Structure (Interactive)}
\usage{
plot_tree_interactive(tree_data, feature_names)
}
\arguments{
\item{tree_data}{Parsed tree JSON data}

\item{feature_names}{Character vector: names for features}
}
\value{
A DiagrammeR object
}
\description{
Create an interactive visualization of tree structure using DiagrammeR.
}
