% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_visualization.R
\name{plot_tree_rules}
\alias{plot_tree_rules}
\title{Plot Tree Rules}
\usage{
plot_tree_rules(tree, X_train, max_rules = 10, ...)
}
\arguments{
\item{tree}{A tree object}

\item{X_train}{Training data for context}

\item{max_rules}{Integer: maximum number of rules to display}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Create a visualization of tree decision rules.
}
