% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tree_viz.R
\name{plot_tree_summary}
\alias{plot_tree_summary}
\title{Create Tree Summary Visualization}
\usage{
plot_tree_summary(model, tree_index = 1)
}
\arguments{
\item{model}{A treefarms model object}

\item{tree_index}{Integer: which tree to summarize}
}
\value{
A ggplot2 object
}
\description{
Create a summary visualization showing tree statistics.
}
