% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.optimaltrees_model}
\alias{predict.optimaltrees_model}
\title{Predict method for optimaltrees_model objects}
\usage{
\method{predict}{optimaltrees_model}(object, newdata, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A optimaltrees_model object}

\item{newdata}{A data.frame or matrix of new features}

\item{type}{Character: "class" or "prob"}

\item{...}{Additional arguments}
}
\value{
Predictions based on type
}
\description{
Predict method for optimaltrees_model objects
}
