% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{predict.optimaltrees_model_loaded}
\alias{predict.optimaltrees_model_loaded}
\title{Predict method for loaded optimaltrees_model objects}
\usage{
\method{predict}{optimaltrees_model_loaded}(object, newdata, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A optimaltrees_model_loaded object}

\item{newdata}{A data.frame or matrix of new features}

\item{type}{Character: "class" or "prob"}

\item{...}{Additional arguments}
}
\value{
Predictions based on type
}
\description{
Predict method for loaded optimaltrees_model objects
}
