% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{predict.treefarms_logloss_model}
\alias{predict.treefarms_logloss_model}
\title{Predict method for treefarms_logloss_model}
\usage{
\method{predict}{treefarms_logloss_model}(object, newdata, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A treefarms_logloss_model object}

\item{newdata}{A data.frame or matrix of new features}

\item{type}{Character string: "class" or "prob"}

\item{...}{Additional arguments (unused)}
}
\description{
Predict method for treefarms_logloss_model
}
