% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.optimaltrees_logloss_model}
\alias{print.optimaltrees_logloss_model}
\title{Print method for optimaltrees_logloss_model}
\usage{
\method{print}{optimaltrees_logloss_model}(x, ...)
}
\arguments{
\item{x}{A optimaltrees_logloss_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for optimaltrees_logloss_model
}
