% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{print.optimaltrees_model}
\alias{print.optimaltrees_model}
\title{Print summary of a TreeFARMS model}
\usage{
\method{print}{optimaltrees_model}(x, ...)
}
\arguments{
\item{x}{A trained TreeFARMS model object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print summary of a TreeFARMS model
}
