% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.treefarms_logloss_model}
\alias{print.treefarms_logloss_model}
\title{Print method for treefarms_logloss_model}
\usage{
\method{print}{treefarms_logloss_model}(x, ...)
}
\arguments{
\item{x}{A treefarms_logloss_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for treefarms_logloss_model
}
