% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_refit.R
\name{reconstruct_recursive}
\alias{reconstruct_recursive}
\title{Reconstruct Tree Recursively}
\usage{
reconstruct_recursive(splits, leaf_preds, path)
}
\arguments{
\item{splits}{List of split information}

\item{leaf_preds}{Named numeric vector}

\item{path}{Current path from root}
}
\value{
Nested list (tree node)
}
\description{
Reconstruct Tree Recursively
}
\keyword{internal}
