% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.optimaltrees_logloss_model}
\alias{summary.optimaltrees_logloss_model}
\title{Summary method for optimaltrees_logloss_model}
\usage{
\method{summary}{optimaltrees_logloss_model}(object, ...)
}
\arguments{
\item{object}{A optimaltrees_logloss_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for optimaltrees_logloss_model
}
