% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.treefarms_logloss_model}
\alias{summary.treefarms_logloss_model}
\title{Summary method for treefarms_logloss_model}
\usage{
\method{summary}{treefarms_logloss_model}(object, ...)
}
\arguments{
\item{object}{A treefarms_logloss_model object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for treefarms_logloss_model
}
