% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefarms.R
\name{summary.treefarms_model}
\alias{summary.treefarms_model}
\title{Summary of a TreeFARMS model}
\usage{
\method{summary}{treefarms_model}(object, ...)
}
\arguments{
\item{object}{A trained TreeFARMS model object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary of a TreeFARMS model
}
