% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon.R
\name{tree_to_json}
\alias{tree_to_json}
\title{Convert Tree to JSON String}
\usage{
tree_to_json(tree)
}
\arguments{
\item{tree}{A TreeClassifier object from the Rashomon set}
}
\value{
A character string containing the tree's JSON representation
}
\description{
Convert a TreeClassifier object to its JSON string representation.
This is used for comparing tree structures.
}
\examples{
\dontrun{
trees <- get_rashomon_trees(model)
tree_json <- tree_to_json(trees[[1]])
}

}
