% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon.R
\name{find_tree_intersection}
\alias{find_tree_intersection}
\title{Find Intersection of Trees Across Rashomon Sets}
\usage{
find_tree_intersection(rashomon_list, verbose = TRUE)
}
\arguments{
\item{rashomon_list}{A list of Rashomon sets, where each element is a list of trees}

\item{verbose}{Logical. Print progress information. Default: TRUE}
}
\value{
A list containing:
  \item{intersecting_trees}{List of tree objects appearing in all Rashomon sets}
  \item{n_intersecting}{Number of intersecting trees}
  \item{tree_jsons}{JSON representations of intersecting trees}
}
\description{
Find trees that appear in ALL Rashomon sets from cross-fitting.
This is the core function for identifying stable trees.
}
\details{
This function compares trees across K Rashomon sets by converting each tree
to its JSON representation and finding common trees across all sets.
}
\examples{
\dontrun{
# Get Rashomon sets from K models
rashomon_sets <- lapply(models, get_rashomon_trees)

# Find trees in all sets
result <- find_tree_intersection(rashomon_sets)
cat("Found", result$n_intersecting, "stable trees\n")
}

}
