% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon.R
\name{get_rashomon_trees}
\alias{get_rashomon_trees}
\title{Extract Trees from Rashomon Set}
\usage{
get_rashomon_trees(model)
}
\arguments{
\item{model}{A trained TreeFARMS model object (class `treefarms_model`)}
}
\value{
A list of tree objects from the model's Rashomon set
}
\description{
Extract all trees from a trained TreeFARMS model's Rashomon set.
}
\examples{
\dontrun{
model <- treefarms(X, y, regularization = 0.1)
trees <- get_rashomon_trees(model)
cat("Number of trees:", length(trees), "\n")
}

}
