% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon_analysis.R, R/tree_visualization.R
\name{plot_tree_rules}
\alias{plot_tree_rules}
\title{Plot Tree Rules}
\usage{
plot_tree_rules(tree, X_train, max_rules = 10, ...)

plot_tree_rules(tree, X_train, max_rules = 10, ...)
}
\arguments{
\item{tree}{A tree object}

\item{X_train}{Training data for context}

\item{max_rules}{Integer: maximum number of rules to display}

\item{...}{Additional arguments}

\item{tree_json}{Character: JSON string representation of a tree}

\item{feature_names}{Character vector: names of features (optional)}

\item{title}{Character: plot title}
}
\value{
A ggplot2 object

A ggplot2 object
}
\description{
Create a visualization of tree rules.

Create a visualization of tree decision rules.
}
\examples{
\dontrun{
# Plot tree rules
plot_tree_rules(tree_json, feature_names = colnames(X))
}

}
