% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cf_rashomon}
\alias{predict.cf_rashomon}
\title{Predict method for cf_rashomon objects}
\usage{
\method{predict}{cf_rashomon}(object, newdata, type = c("class", "prob"), ensemble = TRUE, ...)
}
\arguments{
\item{object}{A cf_rashomon object}

\item{newdata}{A data.frame or matrix of new features}

\item{type}{Character: "class" or "prob"}

\item{ensemble}{Logical: whether to ensemble across all intersecting trees}

\item{...}{Additional arguments}
}
\value{
Predictions based on type
}
\description{
Predict method for cf_rashomon objects
}
