% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_fitted_rashomon.R, R/s3_methods.R
\name{print.cf_rashomon}
\alias{print.cf_rashomon}
\title{Print Method for cf_rashomon Objects}
\usage{
\method{print}{cf_rashomon}(x, ...)

\method{print}{cf_rashomon}(x, ...)
}
\arguments{
\item{x}{A cf_rashomon object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for cf_rashomon Objects

Print method for cf_rashomon
}
