% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R, R/treefarms.R
\name{print.optimaltrees_model}
\alias{print.optimaltrees_model}
\title{Print method for optimaltrees_model}
\usage{
\method{print}{optimaltrees_model}(x, ...)

\method{print}{optimaltrees_model}(x, ...)
}
\arguments{
\item{x}{A trained TreeFARMS model object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for optimaltrees_model

Print summary of a TreeFARMS model
}
