% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_fitted_rashomon.R, R/s3_methods.R
\name{summary.cf_rashomon}
\alias{summary.cf_rashomon}
\title{Summary Method for cf_rashomon Objects}
\usage{
\method{summary}{cf_rashomon}(object, ...)

\method{summary}{cf_rashomon}(object, ...)
}
\arguments{
\item{object}{A cf_rashomon object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for cf_rashomon Objects

Summary method for cf_rashomon
}
